requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
python-dotenv==1.0.1
google-cloud-firestore==2.16.0
google-auth==2.32.0
pytz==2024.1
